python3-prompt-toolkit1: Old version of prompt_toolkit module
OpenHarmony build system relies on this old version 1.x of prompt_toolkit Python
module.

A patch for adding this to meta-openembedded was submitted but rejected,
noting that we should instead try to convince any users of this old version to
upgrade to current (3.0) version. In this case, that is the OpenHarmony
build_lite.git repository.  So for now, we need to carry this recipe here.

We are using version 1.0.18, which is never than the version used currently by
OpenHarmony project, as the old version does not support Python 3.9+ which we
are using.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>