musl: Remove OpenHarmony hack for struct sockaddr_storage
With openharmony-linux-user.patch, we were modifying sys/socket.h header so
that struct sockaddr_storage was coming from linux/socket.h header instead of
being defined directly.

But upstream linux/socket.h does not do that.  It seems like some kind of
Android convention, that have been applied to kernel headers in OpenHarmony,
and as we now are using kernel headers from Oniro, we don't have that.

As part of the reason for using third party from Oniro, we really do want to
use upstream kernel headers, so we need to drop this change.

Note that this difference can be seen as an API change compared to
OpenHarmony.  Code that would be relying on linux/socket.h defining
struct sockaddr_storage will not work without changes.  But the sensible API
is to include sys/socket.h, and in that case, compatibility is preserved.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>