standard-system-coponents: Use HDC from OpenHarmony 3.1
This updates the git revision for the HDC daemon tool to the latest 3.1
release version.

The hdc client tool doesn't compile on the OpenHarmony 3.0 LTS revision
so we use the 3.1 for both client and daemon parts. This can be seen as
the first step towards the 3.1 sources upgrade.

This comes with a modified BUILD.gn (mainly the one from the 3.0 source
tree with a few fixes) and the ohos.build file that is no more used in
the 3.1 tree.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>