standard-system-components: Build HDC daemon
This adds the executable hdcd (HDC Daemon) to the standard system
components and package it for image installation.

It can be run from the QEMU image with the command:
 $ hdcd -t

The -t option indicates that it will accept TCP connection from the
client. As of now, since we don't have the OH system settings in place,
hdcd will choose a default port for the incoming connections that will
be displayed in the console when executed, as:

 [I][2022-04-22 13:45:20.000241][b6f0e6bc][daemon_tcp.cpp:143] TCP listen on port:[45299]

To connect the client, use the following command on the host machine:
 $ ./hdc tconn QEMU_IP:PORT_NUM

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>