oniro-openharmony-linux.conf: Workaround for testimage based recipes
When doing a direct `inherit testimage` in an image recipe, the required host
tools are not setup properly, so tests fails due to missing commands.  Although
the ${HOSTTOOLS} variable is setup with the tools in the recipe, this is not
enough, as runqemu needs to have ${HOSTTOOLS} setup in
configuration (non-recipe) context.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>