qemuarma7: Fix vmalloc size
The vmalloc kernel parameter does not clearly define which unit is default, and
we were only getting 8 MByte vmalloc memory, which was not even enough to
initialize our kernel, with at least PCI support not working due to vmalloc
failure.

Let's try with 128 MByte instead, which does not result in ny vmalloc failures
during boot for now.  64 MByte was not enough...

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>