CI: Add support for OpenHarmony 3.0 build
This change adds support for OpenHarmony 3.0 and 3.1 builds by adding
the OPENHARMONY_VERSION variable to the build matrices.

It is added to conf/auto.conf by the before_script section of the
.bitbake job.

This is handled differently for the fetch job, as in this case the
variable OPENHARMONY_VERSION is set after the .bitbake before_script. So
the before_script appends 'OPENHARMONY_VERSION = ""' to auto.conf and a
sed substitution is used to set it to the correct value.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>