openharmony-standard: Add distro variable OPENHARMONY_VERSION_FULL
This change adds the variable OPENHARMONY_VERSION_FULL to the distro
configuration file. Its default value is set to OPENHARMONY_VERSION.

The idea is to keep the openharmony-standard recipe version number to
OPENHARMONY_VERSION (i.e. Major.minor) and have the variable
OPENHARMONY_VERSION_FULL indicating the patch version.

With the update of the OpenHarmony source repositories to v3.1.1, the
variable OPENHARMONY_VERSION_FULL is set to "3.1.1", through a new
DISTROOVERRIDES entry.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>