CI: Add support for $CI_ONIRO_PARALLEL_BUILD_MAX_JOBS
This introduces support for new variable to be set by gitlab-runner.

It should be set to the max number of parallel jobs to be spawned. It will be
applied to both the number of concurrent bitbake threads/tasks and the number of
parallel jobs in each make/cmake/meson based bitbake task though. So take
care...

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>