CI: Add support for $CI_ONIRO_PARALLEL_BUILD_MAX_LOADAVG
This introduces support for new variable to be set by gitlab-runner.

It should be set to the load average that parallel builds should try to not go
above.  Build tools such as make and ninja is expected to not spawn no jobs
if/when load average is higher than this number.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>