ACTS Align tests list with mandatory set
ACTS are built for all the components used in the given target build. To be
able to compile OpenHarmony 3.0 we need to include more components than
the mandatory list specifies. As a result we have ACTS test suites built
for components we don't use. Moreover, some tests for mandatory components
are missing. This patch aligns the actual set of ACTS test suites with the
mandatory list.

Signed-off-by: Robert Drab <robert.drab@huawei.com>