openharmony-standard-3.0: Add ptest packages
This change adds ptest package definitions for the OpenHarmony
module, unit, and system tests installed in the folder
${libdir}/openharmony-standard/ptest.

Tests that are known to fail are added to the variable
OPENHARMONY_PTEST_IS_BROKEN and thus are not executed by the test image.

All ptest packages are added to openharmony-standard-ptest as runtime
dependencies.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>