openharmony-standard-3.0: Add pre-init systemd service
This oneshot service creates the folder architecture needed by the
OpenHarmony services and components.

All other OpenHarmony services will be started after this one.

These folders are intended to be created by their own systemd service
units, but since they are not all in place yet, this service acts as a
generic placeholder in the meantime.

The service also change /dev/binder file mode to 0666.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>