openharmony-standard: Fix reuse of openharmony-standard sstate in CI
This fixes the problem that have caused us to always rebuilding
openharmony-standard recipe in CI, even if no changes would seem to
require it.

The fix is to rewrite the SRC_URI variable to not include git revisions
but use SRCREV_* variables for that instead, and then setting up
SRCREV_FORMAT appropriately to allow expansion of SRCPV with a meaningful
and correct value.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>