CI: Don't hardcode the external Docker registry hostname
While we still need to use an external Docker registry, we don't have
to hardcode it, and we don't need to push to external git repository
to update it.  By manually setting CI_REGISTRY, CI_REGISTRY_USER,
CI_REGISTRY_PASSWORD and CI_REGISTRY_IMAGE variables, we can build and
push from gitlab.eclipse.org pipeline.

And with this change, anybody forking meta-openharmony will be able to
use the Docker registry included with their GitLab instance if they
want to.

The downside is that forks on gitlab.eclipse.org needs to setup
CI_REGISTRY_IMAGE variable to get working CI.

Signed-off-by: Esben Haabendal <esben.haabendal@huawei.com>