openharmony-standard-3.0: Fix ACTS hap packing
This sets NODE_OPTIONS to pass '--openssl-legacy-provider' to node 17
and fixes the error 'digital envelope routines::unsupported'.
This was resulting in ACTS hap package not contaning the manifest and js
assets files.

Even if NODE_OPTIONS is set in the openharmony-standard recipe, it also
has to be set in the nodejs invocation environment because the recipe
environment is ignored in build_js_assets.py.

Signed-off-by: Thierry Escande <thierry.escande@huawei.com>