openharmony-standard-3.0: add patch to avoid SIGSEGV when running acts
RenderText::PerformLayout() is causing a SIGSEGV for some reasons. 
The error appears to be raised in the Measure method(). This is
causing hap applications to crash, including ACTS test suites.
As a temporary solution we are removing the part of the code that
is raising the segfault. The side effect of this is that texts are 
not properly rendered.

Apply patch to foundation/ace/ace_engine

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>