hicolliecpptest.cpp: Fix HiCollieCppTest to verify process restart
The previous implementation of HiCollieCppTest used the 'ps -A' command
to check if a process had restarted, which was not reliable.
This patch adds the creation and removal of a file
'x_collie_watchdog_test_file' in the child process that is monitored,
and the test cases now verify the presence and removal of this file to
confirm that the process was restarted.
Also, this commit updates the build recipe in
openharmony-standard_3.0.bb to include the newly created patch file.

Signed-off-by: Kareem Zarka <kareem.zarka@huawei.com>