openharmony-standard: add patch to fix segfault problem
Adding and apply ParagraphTxt-ComputeLineBreaks-skip-sigsegv-code.patch

ParagraphTxt::ComputeLineBreaks() is causing a segmentation fault for some 
reasons. This is causing hap applications to crash, including ACTS test suites
and status bar.
As a temporary solution we skip the part of the code that
is raising the segfault. The side effect of this is that texts are 
not properly rendered.

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>