openharmony-standard: add ace_engine-disable-create-component.patch
foundation/ace/ace_engine: disable graphics

The graphics rendering is slowing the system making some ACTS test suites to timeout.
Since we are aiming for certification for headless system, graphics stack 
should not be required.
Disabling graphics by skipping JSView::CreateComponent() execution.
In the future we will remove this patch when we the graphical part will be needed.

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>