CI: Fix path to toolchain/bundle/tools packages
Now that we are building artifacts for multiple machines (qemuarma7 and rpi4 for
now), we obviuosly cannot cheat and just do this type of wildcard matching to
get the desired .sh image.

Instead, let's cheat a bit differently, and hardcode the TUNE_PKGARCH value for
qemuarma7 for now. Later on, when we add testing for rpi4 we have to revisit
this.

Signed-off-by: Esben Haabendal <esben@geanix.com>