openharmony-standard: Make openharmony-preinit live up to it's name
With the removal of all services from the openharmony-preinit script, this is
now a proper oneshot service, and can/must be run before starting the actual
services.

The goal is still to get rid if this script completely.

Anything it does that is actually needed should be moved to the appropriate OH
service unit files.