openharmony-standard: testRdbStoreUpdate0001-await-updatePromise.patch
adding test-xts-acts-testRdbStoreUpdate0001-await-updatePromise.patch and apply to test/xts/acts .

```
The testRdbStoreUpdate0001 testcase which is part of ActsAppdatamagrJsTest
sometimes fails (in our case on raspberrypi4-64). The reason is that
the values in rdbStore are inspected before the update is completed.
The `await updatePromise` command that is clearly needed was commented out
for some unknown reason. Reintroducing the await command to avoid race
conditions.
```

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>