openharmony-standard: appdatamgr-IsColumnNull-initialize-result.patch
add appdatamgr-IsColumnNull-initialize-result.patch and apply to
foundation/distributeddatamgr/appdatamgr

```
testIsColumnNull0003 ACTS testcase sometimes fail because the
result value of IsColumnNull is unpredictable when an error is returned.
This is because AbsSharedResultSet::IsColumnNull does not set any
value to &isNull when there is an error. The result value is
not initialized, producing an unpredictable output of IsColumnNull.
```

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>