Revert "openharmony-standard: Drop patch disabling graphics rendering"
This reverts commit 312c51f4838cc7ce15542261f448b09d7fe56689.

We are seeing multiple different race conditions causing random failures in
ACTS, many of them (maybe all) goes away when we stub out the mostly stub out
JSView::CreateComponent()

As before, this should be dropped again later on when the problems have been
solved.  And is required in order to get graphics support to work.

Signed-off-by: Esben Haabendal <esben@geanix.com>