openharmony-standard-image-tests: Fix ptest ssh timeout in QEMU
SSH host keys are generated by sshdgenkeys.service, which is started by socket
activation on first SSH connection. On QEMU this can take a very long time, even
more than the 300 seconds that is the default timeout when running SSH commands
testimage recipes.

Pre-generated SSH host keys prevents this.

Note, this is only done in the -test image, as it is insecure and should never
be done in images used for other purposes!

Signed-off-by: Esben Haabendal <esben@geanix.com>