openharmony-standard: Configure libsysparam_hal values
This is the strings returned with GetHardwareModel(), GetHardwareProfile() and
so on.  There are more values in there to configure if we want to.

While the hispark boards seems to have some other, more complicated, mechanism
for providing these values, I have chosen to simply configure them directly in
the //base/startup/syspara_lite for now.  KISS.